#75 중복 클래스명 리팩토링 (정산, 알림 서비스)

See merge request s11-fintech-finance-sub1/S11P21C102!73

+++ backend/src/main/java/com/ssafy/moyeobang/common/persistenceentity/order/OrderJpaEntity.java
@@ -47,7 +47,7 @@ public class OrderJpaEntity extends BaseEntity {
     private List<MemberOrderHistoryJpaEntity> memberOrderHistoryJpaEntities = new ArrayList<>();
 
     @Builder
-    public OrderJpaEntity(String title, int amount, WithdrawJpaEntity withdraw) {
+    public OrderJpaEntity(String title, long amount, WithdrawJpaEntity withdraw) {
         this.title = title;
         this.amount = amount;
         this.withdraw = withdraw;
